normie

+++ just_str.c
@@ -4,7 +4,7 @@ void 	just_str(char **str, int just, char c, int LEFT_RIGHT)
 {
 	if (just < 0)
 		just *= -1;
-	if (ft_strlen(*str) >= just)
+	if (ft_strlen(*str) >= (size_t)just)
 		return ;
 	if (LEFT_RIGHT == LEFT)
 		pad_str(str, just - ft_strlen(*str), c, RIGHT);
