+++ prec_str.c
@@ -8,6 +8,6 @@ void 	prec_str(char **str, int prec)
 
 void 	prec_int(char **str, int prec)
 {
-	if (ft_strlen(*str) < prec && prec != -1)
+	if (ft_strlen(*str) < (size_t)prec && prec != -1)
 		pad_str(str, prec - ft_strlen(*str), '0', 0);
 }
